Revert "feat: Deploy Production Phase 1.2"
