+++ package.json
@@ -16,7 +16,6 @@
         "prepare": "husky install"
     },
     "dependencies": {
-        "@headlessui/react": "1.7.15",
         "@heroicons/react": "2.0.18",
         "autoprefixer": "10.4.14",
         "axios": "1.4.0",
