--- src/components/Profile/Register/index.ts
@@ -1,3 +1,0 @@
-export * from './constants';
-export { getFormData } from './getFormData';
-export { RegisterForm } from './RegisterForm';
