--- src/components/font.ts
@@ -1,8 +1,0 @@
-import { IBM_Plex_Sans_Thai } from 'next/font/google';
-
-export const ibmPlexSansThai = IBM_Plex_Sans_Thai({
-    subsets: ['latin'],
-    weight: ['500', '600', '700'],
-    display: 'swap',
-    variable: '--font-ibm',
-});
