+++ src/context/AuthContext.tsx
@@ -7,6 +7,7 @@ import { useRouter } from 'next/router';
 import {
     ReactNode,
     createContext,
+    use,
     useCallback,
     useContext,
     useEffect,
@@ -79,7 +80,7 @@ const AuthProvider = ({ children }: { children: ReactNode }) => {
         switch (router.pathname) {
             case '/':
                 if (user) {
-                    if (alreadyRegistered) router.push('/profile');
+                    if (alreadyRegistered) router.push('/baan-selection');
                     else router.push('/register');
                 }
                 break;
@@ -87,11 +88,9 @@ const AuthProvider = ({ children }: { children: ReactNode }) => {
                 if (!user) {
                     router.push('/');
                 } else if (alreadyRegistered) {
-                    router.push('/profile');
+                    router.push('/baan-selection');
                 }
                 break;
-            case '/wait-baan-selection':
-            case '/profile':
             case '/baan-selection':
                 if (!user) {
                     router.push('/');
