+++ src/dto/userDTO.ts
@@ -9,8 +9,6 @@ export interface UserDTO {
     nickname?: string;
     email?: string;
     phone?: string;
-    emer_phone?: string;
-    emer_relation?: string;
     line_id?: string;
     disease?: string;
     allergy_food?: string;
@@ -22,5 +20,4 @@ export interface UserDTO {
     group_id?: string;
     baan_id?: string;
     is_got_ticket?: boolean;
-    want_bottle?: boolean;
 }
