--- src/pages/profile.tsx
@@ -1,4 +1,0 @@
-// import WaitForBaanselection from "@/components/Profile/WaitForBaan";
-import BaanSelection from '@/components/Profile/BaanSelection';
-
-export default BaanSelection;
