+++ src/types/baan.ts
@@ -10,13 +10,13 @@ export enum BaanSize {
 }
 
 export interface IShortBaan {
-    id: string;
+    id: number;
     name: string;
     imageUrl: string;
-    size: BaanSize;
 }
 
 export interface IBaan extends IShortBaan {
+    size: BaanSize;
     description: string;
     facebook: string;
     facebookUrl: string;
