+++ src/types/user.ts
@@ -2,7 +2,7 @@ export interface IShortUser {
     id: string;
     firstname: string;
     lastname: string;
-    image_url: string;
+    imageUrl: string;
 }
 
 export interface IUser {
@@ -16,8 +16,6 @@ export interface IUser {
     nickname: string;
     email: string;
     phone: string;
-    emerPhone: string;
-    emerRelation: string;
     lineID: string;
     disease: string;
     allergyFood: string;
@@ -29,5 +27,4 @@ export interface IUser {
     groupId: string;
     isGotTicket: boolean;
     baanId?: string;
-    wantbottle: boolean;
 }
