--- src/utils/baan-selection/types.tsx
@@ -1,7 +1,0 @@
-import { BaanSize, IShortBaan } from '@/types/baan';
-
-export interface SelectedBaanRank extends IShortBaan {
-    //Interface for selected baans (3 baans)
-    size: BaanSize;
-    num: number;
-}
