+++ tailwind.config.js
@@ -10,13 +10,9 @@ const config = {
     theme: {
         extend: {
             colors: {
-                purple: {
-                    300: '#6B3BF1',
-                    400: '#393061',
-                },
+                purple: '#393061',
                 pink: {
                     400: '#E95682',
-                    600: '#C03786',
                     800: '#A45BD0',
                 },
                 yellow: '#FFC32A',
@@ -38,9 +34,6 @@ const config = {
                 '4xl': '2rem',
                 '5xl': '3rem',
                 '6xl': '4.5rem',
-            },
-            maxWidth: {
-                xxs: '14rem',
             },
         },
     },
